% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{excel2eprime_example}
\alias{excel2eprime_example}
\title{Get path to excel2eprime example}
\usage{
excel2eprime_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
excel2eprime comes bundled with some example files in its `inst/extdata`
directory. This function make them easy to access.
}
\examples{
excel2eprime_example()
excel2eprime_example("basic.xlsx")
}
