% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_12}
\alias{split_12}
\title{Split sentence of the 1 * 2 experiment design}
\usage{
split_12(path, col_names = TRUE)
}
\arguments{
\item{path}{Path to the file}

\item{col_names}{column names contains the experiment sentence}
}
\value{
tibble
}
\description{
Split sentence of the 1 * 2 experiment design
}
\examples{
split_12("D:\\\\personal\\\\excel2ePrime\\\\R\\\\12.xlsx","A")
}
