% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_222}
\alias{split_222}
\title{Split sentence of the 2 * 2 * 2 experiment design}
\usage{
split_222(path, col_names = TRUE)
}
\arguments{
\item{path}{Path to the file}

\item{col_names}{column names contains the experiment sentence}
}
\value{
tibble
}
\description{
Split sentence of the 2 * 2 * 2 experiment design
}
\examples{
split_222(excel2eprime_example("222.xlsx"),"A")
}
